move renderer to a forked process so it won't block the vscode UI during render
